Break it.
I tried to do a two-digit coefficient and I think that broke it, so I need to fix that.
I'm getting close I think, but it's still not all the way there. I elininated the necessity
of switching the carets for double asterisks, so that's good, though it threw a monkey
wrench into things. Anyway, the debugging line I have breaks it, which I guess is better
than not knowing. Anyway, get that fixed, and I think it will be good. Sorta.